safari-534.53-branch/LayoutTests/fast/events/onunload-clears-onbeforeunload-expected.txt

main frame - has 1 onunload handler(s)
ALERT: before unload
ALERT: unload
If you didn't hit an assert you PASS.
